split action creators in different modules and fixed import in every component
